added account, object, asset, committee, proposal, witness, worker modules in ChainStore
