+++ lib/chain/src/ChainStore/object/index.js
@@ -1,0 +1,4 @@
+export * from "./fetchObject";
+export * from "./getBalanceObjects";
+export * from "./getObject";
+export * from "./updateObject";
